Small test runner improvements (#3761)
* u16 for FunctionTest PARAMETRIZE

* Speed up CB2_TestRunner

* Use DACS to handle illegal instructions

DACS is available in mgba from https://github.com/mgba-emu/mgba/commit/44e074a15e9651481f7f652ac006a7c9d58cbeb9
This is not 0.10.2, but will presumably be available in the next
release. Alternatively, a GBA BIOS could be used.

---------

+++ include/test/test.h
@@ -59,8 +59,8 @@ extern const struct TestRunner gAssumptionsRunner;
 
 struct FunctionTestRunnerState
 {
-    u8 parameters;
-    u8 runParameter;
+    u16 parameters;
+    u16 runParameter;
 };
 
 extern const struct TestRunner gFunctionTestRunner;
